+++ ACadSharp.Tests/TestModels/BlockRecordNode.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace ACadSharp.Tests.TestModels
+{
+	public class BlockRecordNode : TableEntryNode
+	{
+		public List<EntityNode> Entities { get; set; } = new List<EntityNode>();
+	}
+}
