+++ ACadSharp.Tests/TestModels/ColorData.cs
@@ -1,0 +1,17 @@
+﻿namespace ACadSharp.Tests.TestModels
+{
+	public class ColorData
+	{
+		public bool IsByBlock { get; set; }
+
+		public bool IsByLayer { get; set; }
+
+		public byte Green { get; set; }
+
+		public byte Blue { get; set; }
+
+		public byte Red { get; set; }
+
+		public int Index { get; set; }
+	}
+}
